create socks.c and socks.h. Remove -b and -u options. Fix README.md
